Avoid loading Netty classes if it is not used
Fixes #2047
