+++ pom.xml
@@ -786,6 +786,7 @@
               <include>src/main/java/com/rabbitmq/client/PemReader.java</include>
               <include>src/main/java/com/rabbitmq/client/WriteListener.java</include>
               <include>src/main/java/com/rabbitmq/client/impl/NettyFrameHandlerFactory.java</include>
+              <include>src/main/java/com/rabbitmq/client/impl/NettyUtils.java</include>
               <include>src/main/java/com/rabbitmq/client/impl/Environment.java</include>
               <include>src/main/java/com/rabbitmq/client/observation/**/*.java</include>
               <include>src/test/java/com/rabbitmq/client/AmqpClientTestExtension.java</include>
