+++ src/main/java/com/rabbitmq/client/ConnectionFactory.java
@@ -158,7 +158,8 @@ public class ConnectionFactory implements Cloneable {
 
   private boolean netty = false;
   private FrameHandlerFactory frameHandlerFactory;
-  private final NettyConfiguration nettyConf = new NettyConfiguration(this);
+  // lazily created, so the Netty classes it references are loaded only when Netty is activated
+  private NettyConfiguration nettyConf;
 
   private SslContextFactory sslContextFactory;
 
@@ -822,7 +823,7 @@ public class ConnectionFactory implements Cloneable {
   public boolean isSSL() {
     return getSocketFactory() instanceof SSLSocketFactory
         || sslContextFactory != null
-        || this.nettyConf.isTls();
+        || (this.nettyConf != null && this.nettyConf.isTls());
   }
 
   /**
@@ -1860,6 +1861,9 @@ public class ConnectionFactory implements Cloneable {
    */
   public NettyConfiguration netty() {
     useNetty();
+    if (this.nettyConf == null) {
+      this.nettyConf = new NettyConfiguration(this);
+    }
     return this.nettyConf;
   }
